+++ XCDYouTubeKit/XCDYouTubeKit.h
@@ -10,7 +10,7 @@
 #import "XCDYouTubeOperation.h"
 #import "XCDYouTubeVideo.h"
 #import "XCDYouTubeVideoOperation.h"
-
+#import "XCDYouTubeVideoQueryOperation.h"
 
 #if TARGET_OS_IOS || (!defined(TARGET_OS_IOS) && TARGET_OS_IPHONE)
 #import "XCDYouTubeVideoPlayerViewController.h"